changed current for set_current from uint32_t to uint16_t (65535 mA should be enough, but we save some memory)

+++ src/modules/utils/motordrivercontrol/drivers/DRV8711/drv8711.cpp
@@ -97,7 +97,7 @@ void DRV8711DRV::init (uint16_t cs)
     WriteAllRegisters();
 }
 
-void DRV8711DRV::set_current(uint32_t currentma)
+void DRV8711DRV::set_current(uint16_t currentma)
 {
     // derive torque and gain from current
     float c = currentma / 1000.0F; // current in amps
