+++ src/modules/utils/motordrivercontrol/drivers/TMC26X/TMC26X.h
@@ -168,7 +168,7 @@ public:
      * \param current the maximum motor current in mA
      * \sa getCurrent(), getCurrentCurrent()
      */
-    void set_current(unsigned int current);
+    void set_current(uint16_t current);
 
     /*!
      * \brief readout the motor maximum current in mA (1000 is an Amp)
